// src/_helpers/enums/validator/error.code.enum.ts
export enum ErrorCodeEnum {
    AUTH_FAIL = 'AUTH_FAIL',
    INVALID_QUERY_STRING = 'INVALID_QUERY_STRING',
    ENTITY_NOT_FOUND = 'ENTITY_NOT_FOUND',
    ENTITY_CREATION_FAIL = 'ENTITY_CREATION_FAIL',
    NOT_ENOUGH_RIGHTS = 'NOT_ENOUGH_RIGHTS',
    NOT_AUTH = 'NOT_AUTH',
    EMAIL_ALREADY_IN_USE = 'EMAIL_ALREADY_IN_USE',
    SLUG_ALREADY_IN_USE = 'SLUG_ALREADY_IN_USE',
    DOCUMENT_SECTION_NOT_FOUND = 'DOCUMENT_SECTION_NOT_FOUND',
    DOCUMENT_FILE_REQUIRED = 'DOCUMENT_FILE_REQUIRED',
    DOCUMENT_FILE_TOO_LARGE = 'DOCUMENT_FILE_TOO_LARGE',
    DOCUMENT_EXTENSION_NOT_ALLOWED = 'DOCUMENT_EXTENSION_NOT_ALLOWED',
    DOCUMENT_MIME_NOT_ALLOWED = 'DOCUMENT_MIME_NOT_ALLOWED',
    DOCUMENT_SIGNATURE_MISMATCH = 'DOCUMENT_SIGNATURE_MISMATCH',
    DOCUMENT_FILENAME_INVALID = 'DOCUMENT_FILENAME_INVALID',
    DOCUMENT_DUPLICATE_VERSION = 'DOCUMENT_DUPLICATE_VERSION',
    DOCUMENT_STORAGE_UNAVAILABLE = 'DOCUMENT_STORAGE_UNAVAILABLE',
    DOCUMENT_STORAGE_WRITE_FAILED = 'DOCUMENT_STORAGE_WRITE_FAILED',
    DOCUMENT_NOT_FOUND = 'DOCUMENT_NOT_FOUND',
    DOCUMENT_VERSION_NOT_FOUND = 'DOCUMENT_VERSION_NOT_FOUND',
    DOCUMENT_STATUS_INVALID = 'DOCUMENT_STATUS_INVALID',
    DOCUMENT_CURRENT_VERSION_UNAVAILABLE = 'DOCUMENT_CURRENT_VERSION_UNAVAILABLE',
    DOCUMENT_REORDER_INVALID = 'DOCUMENT_REORDER_INVALID',
    DOCUMENT_TITLE_REQUIRED = 'DOCUMENT_TITLE_REQUIRED',
    DOCUMENT_SHARE_LINK_UNAVAILABLE = 'DOCUMENT_SHARE_LINK_UNAVAILABLE',
    DOCUMENT_SHARE_LINK_NOT_FOUND = 'DOCUMENT_SHARE_LINK_NOT_FOUND',
    DOCUMENT_SHARE_LINK_EXPIRY_INVALID = 'DOCUMENT_SHARE_LINK_EXPIRY_INVALID',
    DOCUMENT_PUBLIC_FILE_UNAVAILABLE = 'DOCUMENT_PUBLIC_FILE_UNAVAILABLE',
}